# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Geem CRM & E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: boolean
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /customers/{id}/ledger:
    get:
      operationId: getCustomerLedger
      tags: [customers]
      summary: Customer ledger
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LedgerEntry"

  # --- INVOICES ---
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, mobile, active]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        cnic:
          type: ["string", "null"]
        vehicleNumber:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ledgerBalance:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
        mobile:
          type: string
        cnic:
          type: string
        vehicleNumber:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        cnic:
          type: string
        vehicleNumber:
          type: string
        email:
          type: string
        address:
          type: string
        city:
          type: string
        notes:
          type: string
        active:
          type: boolean
    CustomerListResponse:
      type: object
      required: [customers, total, page, limit]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    LedgerEntry:
      type: object
      required: [id, date, type, description, debit, credit, balance]
      properties:
        id:
          type: integer
        date:
          type: string
        type:
          type: string
        description:
          type: string
        reference:
          type: ["string", "null"]
        debit:
          type: number
        credit:
          type: number
        balance:
          type: number

    # --- INVOICES ---
...